Set up hook for local HQX implementation
Since the standard library `binascii` module is unavailable whenever the
vendored `binhex` would be used, the latter is modified to refer to
`hqx` in the same package instead.

For future-proofing, `hqx` delegates to a `slow_hqx` implementation,
which could be overridden by a C implementation when available.

The function names are also slightly altered to make more sense in the
context of this change. However, the original interface will be
preserved, so that no further change in `binhex.py` will be required.