Add rle_decode / rle_encode (first attempt)
The rle_encode implementation works somewhat differently from the
original C; rather than using numeric indices into the original data
with a nested loop to detect runs, it tries to use a single stateful
loop (remembering how many repeated characters have been seen so far).

It is an open question whether runs of the RUNCHAR byte (0x90) should be
encoded. The specification is ambiguous:
http://fileformats.archiveteam.org/wiki/RLE90
so it's unclear whether a sequence like 0x90 0x00 0x90 0x09 can be used
to encode 10 0x90 bytes, or whether this would represent 0x90 followed
by 9 0x00 bytes, or just what. It appears that the original
standard library decoder interprets this as "10 0x90 bytes", since it
gets the byte to be repeated from the output buffer. However, it also
appears that the original standard library encoder doesn't attempt this
encoding, and will turn runs of 0x90 input into runs of alternating 0x90
and 0x00 bytes.

The code here is not tested, but should preserve that behaviour.