Fix bugs and typos
The code has been informally tested using some sample data.

Trivia: as I discovered by accident, apparently LibreOffice on Linux can
open binhex-encoded PICTs, as long as the file uses a `\n` line ending
after the header (`binhex.py` uses a `\r\r` separator, but manually
editing this produces a result that LibreOffice opens).

The test data appears not to have actually taken advantage of RLE, but
it could be successfully decoded, and re-encoding and re-decoding the
result was a successful round trip.